Move util classes to their own files
