+++ SteamKit2/SteamKit2/Util/MsgUtil.cs
@@ -1,0 +1,97 @@
+﻿/*
+ * This file is subject to the terms and conditions defined in
+ * file 'license.txt', which is part of this source code package.
+ */
+
+namespace SteamKit2
+{
+    /// <summary>
+    /// Contains various utility functions for handling EMsgs.
+    /// </summary>
+    public static class MsgUtil
+    {
+        private const uint ProtoMask = 0x80000000;
+        private const uint EMsgMask = ~ProtoMask;
+
+        /// <summary>
+        /// Strips off the protobuf message flag and returns an EMsg.
+        /// </summary>
+        /// <param name="msg">The message number.</param>
+        /// <returns>The underlying EMsg.</returns>
+        public static EMsg GetMsg( uint msg )
+        {
+            return ( EMsg )( msg & EMsgMask );
+        }
+
+        /// <summary>
+        /// Strips off the protobuf message flag and returns an EMsg.
+        /// </summary>
+        /// <param name="msg">The message number.</param>
+        /// <returns>The underlying EMsg.</returns>
+        public static uint GetGCMsg( uint msg )
+        {
+            return ( msg & EMsgMask );
+        }
+
+        /// <summary>
+        /// Strips off the protobuf message flag and returns an EMsg.
+        /// </summary>
+        /// <param name="msg">The message number.</param>
+        /// <returns>The underlying EMsg.</returns>
+        public static EMsg GetMsg( EMsg msg )
+        {
+            return GetMsg( ( uint )msg );
+        }
+
+        /// <summary>
+        /// Determines whether message is protobuf flagged.
+        /// </summary>
+        /// <param name="msg">The message.</param>
+        /// <returns>
+        ///   <c>true</c> if this message is protobuf flagged; otherwise, <c>false</c>.
+        /// </returns>
+        public static bool IsProtoBuf( uint msg )
+        {
+            return ( msg & ProtoMask ) > 0;
+        }
+
+        /// <summary>
+        /// Determines whether message is protobuf flagged.
+        /// </summary>
+        /// <param name="msg">The message.</param>
+        /// <returns>
+        ///   <c>true</c> if this message is protobuf flagged; otherwise, <c>false</c>.
+        /// </returns>
+        public static bool IsProtoBuf( EMsg msg )
+        {
+            return IsProtoBuf( ( uint )msg );
+        }
+
+        /// <summary>
+        /// Crafts an EMsg, flagging it if required.
+        /// </summary>
+        /// <param name="msg">The EMsg to flag.</param>
+        /// <param name="protobuf">if set to <c>true</c>, the message is protobuf flagged.</param>
+        /// <returns>A crafted EMsg, flagged if requested.</returns>
+        public static EMsg MakeMsg( EMsg msg, bool protobuf = false )
+        {
+            if ( protobuf )
+                return ( EMsg )( ( uint )msg | ProtoMask );
+
+            return msg;
+        }
+        /// <summary>
+        /// Crafts an EMsg, flagging it if required.
+        /// </summary>
+        /// <param name="msg">The EMsg to flag.</param>
+        /// <param name="protobuf">if set to <c>true</c>, the message is protobuf flagged.</param>
+        /// <returns>A crafted EMsg, flagged if requested.</returns>
+        public static uint MakeGCMsg( uint msg, bool protobuf = false )
+        {
+            if ( protobuf )
+                return msg | ProtoMask;
+
+            return msg;
+        }
+    }
+}
