+++ SteamKit2/SteamKit2/Util/Utils.cs
@@ -7,18 +7,30 @@
 
 using System;
 using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Net;
-using System.Net.Sockets;
 using System.Reflection;
 using System.Runtime.InteropServices;
-using System.Text;
-using SteamKit2.Internal;
 
 namespace SteamKit2
 {
     static class Utils
     {
+        /// <summary>
+        /// Performs an Adler32 on the given input
+        /// </summary>
+        public static byte[] AdlerHash( byte[] input )
+        {
+            ArgumentNullException.ThrowIfNull( input );
+
+            uint a = 0, b = 0;
+            for ( int i = 0; i < input.Length; i++ )
+            {
+                a = ( a + input[ i ] ) % 65521;
+                b = ( b + a ) % 65521;
+            }
+
+            return BitConverter.GetBytes( a | ( b << 16 ) );
+        }
+
         public static string EncodeHexString(byte[] input)
         {
             return Convert.ToHexString(input).ToLower();
@@ -142,299 +154,6 @@ namespace SteamKit2
             where T : Attribute
         {
             return (T[])type.GetTypeInfo().GetCustomAttributes( typeof( T ), inherit );
-        }
-    }
-
-    /// <summary>
-    /// Contains various utility functions for dealing with dates.
-    /// </summary>
-    public static class DateUtils
-    {
-        /// <summary>
-        /// Converts a given unix timestamp to a DateTime
-        /// </summary>
-        /// <param name="unixTime">A unix timestamp expressed as seconds since the unix epoch</param>
-        /// <returns>DateTime representation</returns>
-        public static DateTime DateTimeFromUnixTime(ulong unixTime)
-        {
-            return DateTimeOffset.FromUnixTimeSeconds( (long)unixTime ).DateTime;
-        }
-        /// <summary>
-        /// Converts a given DateTime into a unix timestamp representing seconds since the unix epoch.
-        /// </summary>
-        /// <param name="time">DateTime to be expressed</param>
-        /// <returns>64-bit wide representation</returns>
-        public static ulong DateTimeToUnixTime(DateTime time)
-        {
-            return (ulong)new DateTimeOffset( time ).ToUnixTimeSeconds();
-        }
-    }
-
-    /// <summary>
-    /// Contains various utility functions for handling EMsgs.
-    /// </summary>
-    public static class MsgUtil
-    {
-        private const uint ProtoMask = 0x80000000;
-        private const uint EMsgMask = ~ProtoMask;
-
-        /// <summary>
-        /// Strips off the protobuf message flag and returns an EMsg.
-        /// </summary>
-        /// <param name="msg">The message number.</param>
-        /// <returns>The underlying EMsg.</returns>
-        public static EMsg GetMsg( uint msg )
-        {
-            return ( EMsg )( msg & EMsgMask );
-        }
-
-        /// <summary>
-        /// Strips off the protobuf message flag and returns an EMsg.
-        /// </summary>
-        /// <param name="msg">The message number.</param>
-        /// <returns>The underlying EMsg.</returns>
-        public static uint GetGCMsg( uint msg )
-        {
-            return ( msg & EMsgMask );
-        }
-
-        /// <summary>
-        /// Strips off the protobuf message flag and returns an EMsg.
-        /// </summary>
-        /// <param name="msg">The message number.</param>
-        /// <returns>The underlying EMsg.</returns>
-        public static EMsg GetMsg( EMsg msg )
-        {
-            return GetMsg( ( uint )msg );
-        }
-
-        /// <summary>
-        /// Determines whether message is protobuf flagged.
-        /// </summary>
-        /// <param name="msg">The message.</param>
-        /// <returns>
-        ///   <c>true</c> if this message is protobuf flagged; otherwise, <c>false</c>.
-        /// </returns>
-        public static bool IsProtoBuf( uint msg )
-        {
-            return ( msg & ProtoMask ) > 0;
-        }
-
-        /// <summary>
-        /// Determines whether message is protobuf flagged.
-        /// </summary>
-        /// <param name="msg">The message.</param>
-        /// <returns>
-        ///   <c>true</c> if this message is protobuf flagged; otherwise, <c>false</c>.
-        /// </returns>
-        public static bool IsProtoBuf( EMsg msg )
-        {
-            return IsProtoBuf( ( uint )msg );
-        }
-
-        /// <summary>
-        /// Crafts an EMsg, flagging it if required.
-        /// </summary>
-        /// <param name="msg">The EMsg to flag.</param>
-        /// <param name="protobuf">if set to <c>true</c>, the message is protobuf flagged.</param>
-        /// <returns>A crafted EMsg, flagged if requested.</returns>
-        public static EMsg MakeMsg( EMsg msg, bool protobuf = false )
-        {
-            if ( protobuf )
-                return ( EMsg )( ( uint )msg | ProtoMask );
-
-            return msg;
-        }
-        /// <summary>
-        /// Crafts an EMsg, flagging it if required.
-        /// </summary>
-        /// <param name="msg">The EMsg to flag.</param>
-        /// <param name="protobuf">if set to <c>true</c>, the message is protobuf flagged.</param>
-        /// <returns>A crafted EMsg, flagged if requested.</returns>
-        public static uint MakeGCMsg( uint msg, bool protobuf = false )
-        {
-            if ( protobuf )
-                return msg | ProtoMask;
-
-            return msg;
-        }
-    }
-
-    static class WebHelpers
-    {
-        static bool IsUrlSafeChar( char ch )
-        {
-            if ( ( ( ( ch >= 'a' ) && ( ch <= 'z' ) ) || ( ( ch >= 'A' ) && ( ch <= 'Z' ) ) ) || ( ( ch >= '0' ) && ( ch <= '9' ) ) )
-            {
-                return true;
-            }
-
-            return ch switch
-            {
-                '-' or '.' or '_' => true,
-                _ => false,
-            };
-        }
-
-        public static string UrlEncode( string input )
-        {
-            return UrlEncode( Encoding.UTF8.GetBytes( input ) );
-        }
-
-
-        public static string UrlEncode( byte[] input )
-        {
-            StringBuilder encoded = new StringBuilder( input.Length * 2 );
-
-            for ( int i = 0 ; i < input.Length ; i++ )
-            {
-                char inch = ( char )input[ i ];
-
-                if ( IsUrlSafeChar( inch ) )
-                {
-                    encoded.Append( inch );
-                }
-                else if ( inch == ' ' )
-                {
-                    encoded.Append( '+' );
-                }
-                else
-                {
-                    encoded.AppendFormat( "%{0:X2}", input[ i ] );
-                }
-            }
-
-            return encoded.ToString();
-        }
-    }
-
-    static class NetHelpers
-    {
-        public static IPAddress GetLocalIP(Socket activeSocket)
-        {
-            var ipEndPoint = activeSocket.LocalEndPoint as IPEndPoint;
-
-            if ( ipEndPoint == null || ipEndPoint.Address == IPAddress.Any )
-                throw new InvalidOperationException( "Socket not connected" );
-
-            return ipEndPoint.Address;
-        }
-
-        public static IPAddress GetIPAddress( uint ipAddr )
-        {
-            byte[] addrBytes = BitConverter.GetBytes( ipAddr );
-            Array.Reverse( addrBytes );
-
-            return new IPAddress( addrBytes );
-        }
-
-        public static uint GetIPAddressAsUInt( IPAddress ipAddr )
-        {
-            byte[] addrBytes = ipAddr.GetAddressBytes();
-            Array.Reverse( addrBytes );
-
-            return BitConverter.ToUInt32( addrBytes, 0 );
-        }
-
-        public static IPAddress GetIPAddress( this CMsgIPAddress ipAddr )
-        {
-            if ( ipAddr.ShouldSerializev6() )
-            {
-                return new IPAddress( ipAddr.v6 );
-            }
-            else
-            {
-                return GetIPAddress( ipAddr.v4 );
-            }
-        }
-
-        public static CMsgIPAddress GetMsgIPAddress( IPAddress ipAddr )
-        {
-            var msgIpAddress = new CMsgIPAddress();
-            byte[] addrBytes = ipAddr.GetAddressBytes();
-
-            if ( ipAddr.AddressFamily == AddressFamily.InterNetworkV6 )
-            {
-                msgIpAddress.v6 = addrBytes;
-            }
-            else
-            {
-                Array.Reverse( addrBytes );
-
-                msgIpAddress.v4 = BitConverter.ToUInt32( addrBytes, 0 );
-            }
-
-            return msgIpAddress;
-        }
-
-        public static CMsgIPAddress ObfuscatePrivateIP( this CMsgIPAddress msgIpAddress )
-        {
-            var localIp = msgIpAddress;
-
-            if ( localIp.ShouldSerializev6() )
-            {
-                localIp.v6[ 0 ] ^= 0x0D;
-                localIp.v6[ 1 ] ^= 0xF0;
-                localIp.v6[ 2 ] ^= 0xAD;
-                localIp.v6[ 3 ] ^= 0xBA;
-
-                localIp.v6[ 4 ] ^= 0x0D;
-                localIp.v6[ 5 ] ^= 0xF0;
-                localIp.v6[ 6 ] ^= 0xAD;
-                localIp.v6[ 7 ] ^= 0xBA;
-
-                localIp.v6[ 8 ] ^= 0x0D;
-                localIp.v6[ 9 ] ^= 0xF0;
-                localIp.v6[ 10 ] ^= 0xAD;
-                localIp.v6[ 11 ] ^= 0xBA;
-
-                localIp.v6[ 12 ] ^= 0x0D;
-                localIp.v6[ 13 ] ^= 0xF0;
-                localIp.v6[ 14 ] ^= 0xAD;
-                localIp.v6[ 15 ] ^= 0xBA;
-            }
-            else
-            {
-                localIp.v4 ^= MsgClientLogon.ObfuscationMask;
-            }
-
-            return localIp;
-        }
-
-        public static bool TryParseIPEndPoint( string stringValue, [NotNullWhen( true )] out IPEndPoint? endPoint )
-        {
-            var colonPosition = stringValue.LastIndexOf( ':' );
-
-            if ( colonPosition == -1 )
-            {
-                endPoint = null;
-                return false;
-            }
-
-            if ( !IPAddress.TryParse( stringValue.AsSpan( 0, colonPosition ), out var address ) )
-            {
-                endPoint = null;
-                return false;
-            }
-
-            if ( !ushort.TryParse( stringValue.AsSpan( colonPosition + 1 ), out var port ) )
-            {
-                endPoint = null;
-                return false;
-            }
-
-            endPoint = new IPEndPoint( address, port );
-            return true;
-        }
-
-        public static (string host, int port) ExtractEndpointHost( EndPoint endPoint )
-        {
-            return endPoint switch
-            {
-                IPEndPoint ipep => (ipep.Address.ToString(), ipep.Port),
-                DnsEndPoint dns => (dns.Host, dns.Port),
-                _ => throw new InvalidOperationException( "Unknown endpoint type." ),
-            };
         }
     }
 }
