+++ SteamKit2/SteamKit2/Util/WebHelpers.cs
@@ -1,0 +1,59 @@
+﻿/*
+ * This file is subject to the terms and conditions defined in
+ * file 'license.txt', which is part of this source code package.
+ */
+
+
+
+using System.Text;
+
+namespace SteamKit2
+{
+    static class WebHelpers
+    {
+        static bool IsUrlSafeChar( char ch )
+        {
+            if ( ( ( ( ch >= 'a' ) && ( ch <= 'z' ) ) || ( ( ch >= 'A' ) && ( ch <= 'Z' ) ) ) || ( ( ch >= '0' ) && ( ch <= '9' ) ) )
+            {
+                return true;
+            }
+
+            return ch switch
+            {
+                '-' or '.' or '_' => true,
+                _ => false,
+            };
+        }
+
+        public static string UrlEncode( string input )
+        {
+            return UrlEncode( Encoding.UTF8.GetBytes( input ) );
+        }
+
+
+        public static string UrlEncode( byte[] input )
+        {
+            StringBuilder encoded = new StringBuilder( input.Length * 2 );
+
+            for ( int i = 0 ; i < input.Length ; i++ )
+            {
+                char inch = ( char )input[ i ];
+
+                if ( IsUrlSafeChar( inch ) )
+                {
+                    encoded.Append( inch );
+                }
+                else if ( inch == ' ' )
+                {
+                    encoded.Append( '+' );
+                }
+                else
+                {
+                    encoded.AppendFormat( "%{0:X2}", input[ i ] );
+                }
+            }
+
+            return encoded.ToString();
+        }
+    }
+}
